requirements.txt
Flask
flask-cors
pandas
tensorflow
pillow